+++ python/tests/people/personindex_test.py
@@ -85,6 +85,44 @@ def test_add_person(index_stub):
         index.add_person(Person("yang-liu", index))
 
 
+def test_remove_person(index):
+    # Preconditions
+    pid = "xu-huang-nanjing"
+    name = Name("Xu", "Huang")
+    person = index[pid]
+    assert index._by_orcid.get("0009-0006-0385-4054") == pid
+    assert pid in index._by_name[name]
+    assert pid in index._slugs_to_verified_ids["xu-huang"]
+    assert len(person.item_ids) == 0
+
+    # Remove person
+    index.remove_person(person)
+
+    assert pid not in index
+    assert index._by_orcid.get("0009-0006-0385-4054") is None
+    assert pid not in index._by_name[name]
+    assert pid not in index._slugs_to_verified_ids["xu-huang"]
+
+    with pytest.raises(ValueError):
+        # Can't remove again...
+        index.remove_person(person)
+
+
+def test_remove_person_should_raise(index):
+    assert "yang-liu/unverified" in index
+    person1 = index["yang-liu/unverified"]
+    # Can't remove unverified
+    with pytest.raises(ValueError):
+        index.remove_person(person1)
+
+    assert "yang-liu-icsi" in index
+    person2 = index["yang-liu-icsi"]
+    assert len(person2.item_ids) > 0
+    # Can't remove if papers still attached to Person
+    with pytest.raises(ValueError):
+        index.remove_person(person2)
+
+
 def test_similar_names_defined_in_people_index(index_stub):
     index = index_stub
     index.reset()
@@ -737,9 +775,33 @@ def test_namespec_change_name_affects_name_resolution(anthology):
     namespec.name = Name("Nathan Middlename", "Noiry")
     person2 = namespec.resolve()
     assert person2 is not person1
+    assert person2.id == UNVERIFIED_PID_FORMAT.format(pid="nathan-middlename-noiry")
     assert item_id not in person1.item_ids
     assert item_id in person2.item_ids
-    assert person2.id == UNVERIFIED_PID_FORMAT.format(pid="nathan-middlename-noiry")
+
+
+def test_namespec_change_name_affects_volume_and_frontmatter(anthology):
+    index = anthology.people
+    # Precondition: Find a volume that resolves to a given (unverified) person
+    item_id = ("2022.acl", "long", None)
+    frontmatter_id = ("2022.acl", "long", "0")
+    namespec = anthology.get_volume(item_id).editors[-1]
+    person1 = index.get(UNVERIFIED_PID_FORMAT.format(pid="aline-villavicencio"))
+    assert namespec.resolve() is person1
+    assert item_id in person1.item_ids
+    assert frontmatter_id in person1.item_ids
+
+    # Changing the name should move the volume AND its frontmatter
+    namespec.name = Name("Aline Middlename", "Villavicencio")
+    person2 = namespec.resolve()
+    assert person2 is not person1
+    assert person2.id == UNVERIFIED_PID_FORMAT.format(
+        pid="aline-middlename-villavicencio"
+    )
+    assert item_id not in person1.item_ids
+    assert frontmatter_id not in person1.item_ids
+    assert item_id in person2.item_ids
+    assert frontmatter_id in person2.item_ids
 
 
 def test_namespec_change_id_affects_name_resolution(anthology):
@@ -758,6 +820,28 @@ def test_namespec_change_id_affects_name_resolution(anthology):
     assert namespec.resolve() is person2
     assert item_id not in person1.item_ids
     assert item_id in person2.item_ids
+
+
+def test_namespec_change_id_affects_volume_and_frontmatter(anthology):
+    index = anthology.people
+    # Precondition: Find a volume that resolves to a given (unverified) person
+    item_id = ("2022.acl", "long", None)
+    frontmatter_id = ("2022.acl", "long", "0")
+    namespec = anthology.get_volume(item_id).editors[-1]
+    person1 = index.get(UNVERIFIED_PID_FORMAT.format(pid="aline-villavicencio"))
+    assert namespec.resolve() is person1
+    assert item_id in person1.item_ids
+    assert frontmatter_id in person1.item_ids
+
+    # Changing the name should move the volume AND its frontmatter
+    namespec.id = "aline-villavicencio-test"
+    person2 = namespec.resolve()
+    assert person2 is not person1
+    assert person2.id == "aline-villavicencio-test"
+    assert item_id not in person1.item_ids
+    assert frontmatter_id not in person1.item_ids
+    assert item_id in person2.item_ids
+    assert frontmatter_id in person2.item_ids
 
 
 def test_namespec_remove_id_affects_name_resolution(anthology):